requirements.txt
absl-py==2.1.0
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
astunparse==1.6.3
attrs==24.3.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dlib==19.24.6
exceptiongroup==1.2.2
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.6
flatbuffers==24.12.23
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.69.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pytz==2024.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
